ДЗ Calculator

+++ src/main/java/pro/sky/calculator/CalculatorServiceImpl.java
@@ -10,17 +10,17 @@ public class CalculatorServiceImpl implements CalculatorService {
 
     public String functionPlusInCalculator(int firstParam, int secondParam) {
 //        if (firstParam == null || secondParam == null)
-        int result = firstParam + secondParam;
+        float result = firstParam + secondParam;
         return firstParam + " + " + secondParam + " = " + result;
     }
 
     public String functionMinusInCalculator(int firstParam, int secondParam) {
-        int result = firstParam - secondParam;
+        float result = firstParam - secondParam;
         return firstParam + " - " + secondParam + " = " + result;
     }
 
     public String functionMultiplyInCalculator(int firstParam, int secondParam) {
-        int result = firstParam * secondParam;
+        float result = firstParam * secondParam;
         return firstParam + " * " + secondParam + " = " + result;
     }
 
@@ -28,7 +28,7 @@ public class CalculatorServiceImpl implements CalculatorService {
         if (secondParam == 0) {
             return "Деление на ноль невозможно!";
         }else {
-            int result = firstParam / secondParam;
+            float result = firstParam / secondParam;
             return firstParam + " / " + secondParam + " = " + result;
         }
     }
